backend/requirements.txt
Flask==2.3.4
flask-cors==4.0.0
pymongo==4.6.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.0
